Update NYC slug value to match Hub
In Hub, the New York City location uses a slug value of "new-york-city".

This was not always this way, and it's possible that the "new-york" value used
previously in the location service did correctly match the Hub counter part at
some point in time.

That said, the mismatch is causing the hiring form to never auto-populate the
NYC option when the IP detection returns that location. This change correct the
mismatch.